clean up types.

+++ main.py
@@ -30,7 +30,7 @@ from protorpc import remote
 
 # print 'file', endpoints.__file__
 
-class GreetingEP(endpoints_model.EndpointsModel):
+class Greeting(endpoints_model.EndpointsModel):
     message = ndb.StringProperty()
 
     # Define IdSet() and id() to require a string id.
@@ -45,32 +45,16 @@ class GreetingEP(endpoints_model.EndpointsModel):
         return self.key.string_id()
 
 
-class GreetingCollectionEP(endpoints_model.EndpointsModel):
-    items = ndb.StructuredProperty(GreetingEP, repeated=True)
-
-
-
 ####################################################################
 # TEMPORARY
 class Result(messages.Message):
     message = messages.StringField(1)
 
 
-class Greeting(messages.Message):
-    """Greeting that stores a message."""
-    message = messages.StringField(1)
-    id = messages.StringField(2)
-
-
-class GreetingCollection(messages.Message):
-    """Collection of Greetings."""
-    items = messages.MessageField(Greeting, 1, repeated=True)
-
-
-STORED_GREETINGS = GreetingCollectionEP(items=[
-    GreetingEP(message='hello world!', id='banana'),
-    GreetingEP(message='goodbye world!', id='orange'),
-])
+STORED_GREETINGS = [
+    Greeting(message='hello world!', id='banana'),
+    Greeting(message='goodbye world!', id='orange'),
+]
 ####################################################################
 
 
@@ -81,6 +65,10 @@ STORED_GREETINGS = GreetingCollectionEP(items=[
 WEB_CLIENT_ID = '294885104230-a8i6fnv7pcgsg0chm0r8vtihcftkcurj.apps.googleusercontent.com'
 PS_CLIENT_ID = '294885104230-ue8r3k8on02m0kjsu8vh5ulnke7imnrf.apps.googleusercontent.com'
 ADMIN_USERS = ['soltesz@google.com', 'stephen.soltesz@gmail.com']
+
+ALLOWED_CLIENT_IDS = [
+    WEB_CLIENT_ID, PS_CLIENT_ID, endpoints.API_EXPLORER_CLIENT_ID]
+
 
 def authorized(f):
     @functools.wraps(f)
@@ -96,19 +84,17 @@ def authorized(f):
 
 
 @endpoints.api(name='greeting', version='v1',
-               allowed_client_ids=[
-                   WEB_CLIENT_ID, PS_CLIENT_ID, endpoints.API_EXPLORER_CLIENT_ID])
+               allowed_client_ids=ALLOWED_CLIENT_IDS)
 class GreetingApi(remote.Service):
 
-    @GreetingEP.query_method(query_fields=('limit', 'order', 'pageToken'),
-                             path='greetings2')
+    @Greeting.query_method(query_fields=('limit', 'order', 'pageToken'),
+                           path='greetings2')
     @authorized
     def list(self, query):
         logging.debug('query: %s', query)
         return query
 
-    @GreetingEP.method(
-        path='greetings2/{id}', http_method='GET')
+    @Greeting.method(path='greetings2/{id}', http_method='GET')
     @authorized
     def get(self, greeting):
         if not greeting.from_datastore:
@@ -116,31 +102,29 @@ class GreetingApi(remote.Service):
                     'Greeting not found: "%s"' % greeting.key.id())
         return greeting
 
-    @GreetingEP.method(path='greetings2', http_method='POST')
+    @Greeting.method(path='greetings2', http_method='POST')
     @authorized
     def create(self, greeting):
         greeting.put()
         return greeting
 
-    @GreetingEP.method(
-        response_fields=(),
-        path='greetings2/{id}', http_method='POST')
+    @Greeting.method(
+        response_fields=(), path='greetings2/{id}', http_method='POST')
     @authorized
     def delete(self, greeting):
+        if not greeting.from_datastore:
+            raise endpoints.NotFoundException(
+                    'Greeting not found: "%s"' % greeting.key.id())
         greeting.key.delete()
         return greeting
 
 
     # TEMPORARY
     @endpoints.method(
-        message_types.VoidMessage, Result,
-        path='setup2', http_method='GET')
+        message_types.VoidMessage, Result, path='setup2', http_method='GET')
     @authorized
     def setup2(self, unused_request):
-        for greeting in STORED_GREETINGS.items:
-            #g = GreetingEP(message=greeting.message)
-            #g.key = ndb.Key(GreetingEP, greeting.id)
-            #g.put()
+        for greeting in STORED_GREETINGS:
             greeting.put()
         return Result(message='okay')
 
@@ -148,13 +132,7 @@ class GreetingApi(remote.Service):
         message_types.VoidMessage, Result,
         path='authcheck', http_method='GET')
     def authcheck(self, unused_request):
-        # print dir(self)
-        # print unused_request
         user = endpoints.get_current_user()
-        # request_path = urlparse.urlsplit(self.request.url).path
-        # if request_path.startswith('/_ah/api/'):
-        #     logging.info('path: %s', request_path)
-
         if not user:
             raise endpoints.ForbiddenException(
                 'Users must be authenticated!')
@@ -167,67 +145,6 @@ class GreetingApi(remote.Service):
         return Result(message=msg)
 
 
-    # [START multiply]
-    # This ResourceContainer is similar to the one used for get_greeting, but
-    # this one also contains a request body in the form of a Greeting message.
-#    MULTIPLY_RESOURCE = endpoints.ResourceContainer(
-#        Greeting,
-#        times=messages.IntegerField(2, variant=messages.Variant.INT32,
-#                                    required=True))
-#
-#    @endpoints.method(
-#        # This method accepts a request body containing a Greeting message
-#        # and a URL parameter specifying how many times to multiply the
-#        # message.
-#        MULTIPLY_RESOURCE,
-#        # This method returns a Greeting message.
-#        Greeting,
-#        path='greetings/multiply/{times}',
-#        http_method='POST',
-#        name='greetings.multiply')
-#    def multiply_greeting(self, request):
-#        return Greeting(message=request.message * request.times)
-#    # [END multiply]
 
-
-# [START auth_config]
-WEB_CLIENT_ID = '294885104230-a8i6fnv7pcgsg0chm0r8vtihcftkcurj.apps.googleusercontent.com'
-# WEB_CLIENT_ID = 'replace this with your web client application ID'
-ANDROID_CLIENT_ID = 'replace this with your Android client ID'
-IOS_CLIENT_ID = 'replace this with your iOS client ID'
-ANDROID_AUDIENCE = WEB_CLIENT_ID
-ALLOWED_CLIENT_IDS = [
-    WEB_CLIENT_ID, ANDROID_CLIENT_ID, IOS_CLIENT_ID,
-    endpoints.API_EXPLORER_CLIENT_ID]
-# [END auth_config]
-
-
-# [START authed_greeting_api]
-#@endpoints.api(
-#    name='authed_greeting',
-#    version='v1',
-#    # Only allowed configured Client IDs to access this API.
-#    allowed_client_ids=ALLOWED_CLIENT_IDS,
-#    # Only allow auth tokens with the given audience to access this API.
-#    audiences=[ANDROID_AUDIENCE],
-#    # Require auth tokens to have the following scopes to access this API.
-#    scopes=[endpoints.EMAIL_SCOPE])
-#class AuthedGreetingApi(remote.Service):
-#
-#    @endpoints.method(
-#        message_types.VoidMessage,
-#        Greeting,
-#        path='greet',
-#        http_method='POST',
-#        name='greet')
-#    def greet(self, request):
-#        user = endpoints.get_current_user()
-#        user_name = user.email() if user else 'Anonymous'
-#        return Greeting(message='Hello, {}'.format(user_name))
-## [END authed_greeting_api]
-
-
-# [START api_server]
-# api = endpoints.api_server([GreetingApi, AuthedGreetingApi])
+# The api server.
 api = endpoints.api_server([GreetingApi])
-# [END api_server]
